requirements.txt
pika
python-daemon
lockfile
